support poetry terminal activation required for versions > 2.0.0
